Tidying up Typescript and file structure.
