+++ src/Elsa.Dashboard/src/models/custom-component-models.js
@@ -1,0 +1,2 @@
+//import { x } from '@elsa-workflows/elsa-workflows-studio'; //Not In Use.  Using as a placeholder to prompt developers to draw, and copy any interfaces they need from Elsa directly.
+export {};
