+++ src/Elsa.Dashboard/src/models/utils.ts
@@ -1,0 +1,3 @@
+export type Map<T> = {
+  [key: string]: T
+};
